Update patch set 4
Patch Set 4:

> > But please add the bug footer (bug 536324), and see Dani's
 > comment
 > > there. Does using o.e.j.api.errors.CanceledException force us to
 > do
 > > a major version increase?
 > 
 > yes, you are right this breaks API since CanceledException is a
 > checked exception these methods didn't throw so far. We can use
 > CancelledException to avoid this since it is a subclass of
 > IOException which was thrown by the affected public classes before:
 > https://git.eclipse.org/r/#/c/127172/

Yes, that's the work-around for now. Too bad; and sorry I didn't think of this sooner. PDE's API checks also don't warn about it...

The problem now is of course that existing users of the rename detector may have to handle the CancelledException explicitly without getting compiler help to notify them of the change...

So we'll have to clean this up when we tackle the "too many exceptions derived from IOException" issue.

Patch-set: 4
Label: Verified=0