Update patch set 1
Patch Set 1: Code-Review+1

The change looks okay. I was just wondering if the sets of conflicts are actually sorted (it looks like that in the screenshot, but might be just by chance).

And a side note: I've never seen more than one entry in that dialog on Windows, so this behavior of printing this as non-readable character seems platform-dependent.

Patch-set: 1
Reviewer: Gerrit User 440 <440@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1