Update patch set 5
Patch Set 5:

(2 comments)

Looks not bad, but needs more work when several repositories are selected. I would also like to have some indication in the menu that more than one repo will be switched. Maybe change the menu text to "Switch repositories to". (Or maybe you have a better idea.)

Note that I took the liberty of rebasing these two changes, so before you continue working on this, fetch them from Gerrit and then continue the work on that.

Patch-set: 5