Update patch set 5
Patch Set 5:

Thanks for the update. I'll be sure to grab the newest copy when I get to work on it again.

I agree wholeheartedly with all that you said. 
- Changing the labels as a way to indicate that the multi selection is in place is a really good idea.
- I have to agree also that the behavior you mentioned about multiple jobs being spawned does seem a bit off whenever the checkouts don't run cleanly.

With these considerations in mind, the first thing that comes to my head is the view implemented for the Pull action. The multi-repository pull operation is considered as a single job for eclipse and the results are grouped in a "Pull Result for Multiple Repositories" window, where each line represents a repository and its status. Do you think an approach similar to that would work? The only thing I might change would be to not display this result dialog if there are no conflicts, sparing one click (maybe it is indeed better do display it regardless for consistency sake, I don't know). For that to work, we may have to figure out a way to emulate a dry run on a checkout command so that the whole multicheckout operation is atomic. What do you think?

Patch-set: 5