Update patch set 5
Patch Set 5:

> Do you think an approach similar to that would work?

Yes, I think we would need something similar. Then we can do the "running launches" check once before, and a refresh once at the end. The project tracker
(which takes care of handling projects appearing/disappearing due to the branch switch) is another thing to keep an eye on. Perhaps that also needs to be re-done
such that it can create/remove projects only once for all repos.

 > emulate a dry run on a checkout command

JGit's DirCacheCheckout already is implemented as a two-phase algorithm that first figures out what to do, and only then does the real modifications if there are no conflicts. So this seems doable with not too much work.

But I'd do this in several steps. First do the EGit-only changes (doing it in one job: pre-conditions, checking-out, refresh); if some repos couldn't be checked out, display a dialog with the repos that couldn't be switched.

Then do the dry-run functionality in JGit, and then finally use it in EGit.

But that's just a suggestion. If you see another way to proceed it's fine, too.

Patch-set: 5