Update patch set 1
Patch Set 1:

> However, I don't find a way to realize your suggested solution,

It is doable. Just gave it a try, but as usual one runs into arcane bugs in Eclipse. Concerning the actions, their enablement, the display of the icons (always shown as "enabled"), and the presence of the commands in the quick access (only when really enabled), it works.

* Adding actions programmatically at the end of createPartControl() adds the refresh and "link with selection" actions at the front of the toolbar, but otherwise works fine.
* It is possible to delay adding the actions until the framework has added the contributions from plugin.xml and then add the actions at the place where they currently are. As far as the actions are concerned, that also works. But the framework has a bug and gives (at least on OS X, Neon) the toolbar a gray background if it cannot be shown right of the rightmost tab but is moved below the tabs.

The last time I hit this was when I did the find toolbar in the git history view. There I was able to figure out a super-hacky work-around that appears to work most of the time, but here I don't see any.

(And no, I haven't yet reported the bug. Would need to create a little stand-alone plug-in to demonstrate the problem first. And would have to check on Photon first. Also don't know whether it's a renderer or a CSS problem.)

 > Thomas, would you be okay with just preventing the exception?

So yes, for the "link with selection" toolbar button leaving the command always enabled and fixing the code is a solution. The only drawback is that it'll appear in quick access when it shouldn't. I think, though, the action should have a tooltip then (or whatever it is that quick access displays as description). If it can show in quick access when the repositories view is not enabled, the user needs some explanation that this command relates to the git repositories view.

Note that the "Refresh" button has a similar problem (shown disabled when current active part is one that doesn't support refresh), but there always enabling our handler won't work (will lead to handler conflicts).

Patch-set: 1