Update patch set 5
Patch Set 5:

That sounds fun!

I started working on the first change where the label for the 'Switch To' menu should be changed to 'Switch repositories to...' as an additional visual cue to the user that he will be switching multiple repositories at the same time. I couldn't manage to find a working solution, though. The path that I followed was looking into the PushMenu.java implementation that has a label that is updated according to the currently checked out branch. However, I don't see how/where I could add submenu items.

<((1)) dynamic
      id="egit.menu.switch"
      label="%SwitchToMenu.label"
      icon="icons/obj16/branches_obj.png"
      class="org.eclipse.egit.ui.internal.actions.PushMenu2">
   <dynamic
         class="org.eclipse.egit.ui.internal.actions.SwitchToMenu"
         id="org.eclipse.egit.ui.switchToMenu">
   </dynamic>
   <visibleWhen
         checkEnabled="false">
      <test
            property="GitSelection.resourcesAllInRepository">
      </test>
   </visibleWhen>
</dynamic>

Changing the ((1)) tag to dynamic allowed me to intercept the creation of the menu and properly update the label, however, that also disabled the nested ((2)) dynamic, or so it seems. I believe that once there is a dynamic tag, anything nested inside of it should be declared programmatically as well? I tried searching the internet and came across some old posts which mentioned AbstractContributionFactory in conjunction with IMenuService, but couldn't make sense out of it. Do you have some guidelines on I how I can achieve this? 

Thanks for the help.

Patch-set: 5
Label: Verified=0