Update patch set 5
Patch Set 5:

> I started working on the first change where the label for the
 > 'Switch To' menu should be changed to 'Switch repositories to...'

The easiest way to achieve this is probably to have the command in the menu in plugin.xml twice, with two different labels, and <visibleWhen> clauses that enable only one of them: one is shown when the selection is all in one repository, the other when the selection covers multiple repositories.

Check what I did for the org.eclipse.egit.ui.history.CheckoutCommand. If the commit has a single ref (branch), it shows "Check Out". When there are multiple branches on the commit, it shows "Check Out..." (because the user will then get a dialog asking _which_ branch should be checked out).

Patch-set: 5
Label: Verified=0