Update patch set 1
Patch Set 1:

Do we know _why_ these exceptions occur? To me, this looks like a bug in platform, which we mask here only. If I read the stack trace in the bug correctly, platform is trying to _create_ a preference node when the parent node has already been removed, and is doing so during a refreshLocal() on the project? That should just never occur!

Patch-set: 1