Update patch set 7
Patch Set 7:

Thanks for the tip!

This is the first ministep for the implementation, which is to have the label change according to the number of repositories selected. I ended up having to create another property tester because none of the others guaranteed a multi-repository selection.

While testing, I actually notice that maybe the property tester for 'projectsSingleRepository' could be improved. Because if I select a project in the project explorer and one random folder from another project, the 'Switch To' option is still visible. I was thinking that maybe I could fix this with another patch so as to not over clutter this one. What do you think?

Is jenkins stable? It is notifying me of failure but it appears to not be related to code and more to infrastructure.

Patch-set: 7
Label: Verified=0