Update patch set 1
Patch Set 1:

> Are you sure about the _displayed_ part? AFAIK labels are
 > computed for all elements, whether visible or not.

Just verified this (without this change). Started with an empty reflog view with 6 rows visible. Loading the reflog of my EGit clone, which has 1515 lines, produced 1515 calls to the label provider. The framework computes labels for all items that are (logically) visible, whether they are displayed (shown on the screen) or not. (So for children of a non-expanded tree node, labels will be computed when the node is expanded.) But displaying 1515 items in a table with only 6 rows computes all 1515 labels up front, even those for items not shown in these 6 rows.

So computing these 1515 labels up front in the background, using a single walk, is definitely no worse that using 1515 separate walks. And we can omit the setRetainBody.

Patch-set: 1