Fix layout and title in CommitSelectionDialog
Since commit 7011de7 the Table of a CommitGraphTable is wrapped
in an intermediate Composite for its TableColumnLayout so that
column widths can be adapted in response to resizes.

However, CommitGraphTable.getControl() still returned the table.
As a result, the GridLayoutData used in the CommitSelectionDialog
got attached to the wrong widget.

Fix this by returning the intermediate Composite from getControl().
Add a new getTable() method that indeed returns the Table; it's
needed for focus tracking. (The getControl().setFocus() calls are
fine; the Composite forwards the focus to its first focusable child,
which is the Table.)

Also use the short repository name in the dialog title instead of
the git directory path.

Bug: 537959
Change-Id: I636982330ca53888d1a23e3ba1ff1f8698829de8
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>