[WIP] Adds repository selection combo to the fetch gerrit change dialog
Removes the message that no repository is selected if the fetch gerrit
change dialog is called without a repository selected (e.g. via ctrl +
3).
Instead there is a combo selector on top of the dialog that lets the
user choose what repository he wants (even if a repository is selected).

TODO:
- Message on top of dialog is not synced with the current selection
- ...

Bug: 507733
Change-Id: I1b8adb4ecf1b769d249f1d7180d6f4a0473131f4
Signed-off-by: Jonas Hungershausen <jonas.hungershausen@vogella.com>