Update patch set 4
Patch Set 4:

(1 comment)

Build fails because this is based on a too old parent. We had a release in the meantime.

This needs a rebase onto current origin/master. Since you could fix the commit message in that step, too, I'll leave that to you guys. Once it builds I'll take a look.

Patch-set: 4
Label: Code-Review=0