Update patch set 8
Patch Set 8: Code-Review-1

(3 comments)

Sorry, I don't understand this change at all. What problem is this trying to solve?

The original request in bug 507733 was to be able to choose a repository if none was selected. Period. That does not require any fancy active editor stuff (which may be on a file not in any git repo anyway). And FetchChangeFromGerritCommand doesn't even open this wizard if it cannot detect a repo.

To satisfy the request in bug 507733, all that's needed is in FetchChangeFromGerritCommand, if repository == null, to show a simple dialog where the user can select one of the repositories shown in the repositories view. Once the user has chosen one, open the FetchFromGerritWizard with that repo. If the user canceled, quit. So basically replace the MessageDialog in FetchChangeFromGerritCommand with a new and better dialog.

Patch-set: 8
Label: Code-Review=-1