Update patch set 8
Patch Set 8:

> I think that if no repository is selected the active editor is a
 > reasonable choice to guess the repository to take as default. This
 > assumes that I tend to look at files which are related to the
 > change I plan to pull in.

One can maybe look at the active editor to try to determine a preselection for the repository selection dialog that would be used instead of the plain message dialog.  Such an approach would also have the benefit that it would be much easier to implement.

The current combo approach breaks some existing functionality and, because it always goes over all configured repos, is bound to give poor performance. I don't like it.

A repo selection on the FetchGerritChangePage makes no sense. If you've already entered a change ref, it's highly unlikely that it will make sense if you then change the repo. Let's stick with a simple workflow: first choose a repo (if there's none that can be determined from the event), then open the fetch wizard. FetchGerritChangeWizard and FetchGerritChangePage would remain completely unchanged with that approach.

Patch-set: 8