Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Doesn't match the indentation in the table header. See inline comment on lines 529/537 for another approach: only custom-draw column 1. That gives a default indentation that is consistent with the table header.

On Mac we also get the ellipses when the column isn't wide enough to show the whole text.

Tested on Mac and on Linux/gtk 3.22.10 (which doesn't do the ellipses bit, though). Please double-check on Windows.

Screenshot attached to bug 539553.

Patch-set: 1
Label: Code-Review=-1