Indent history table columns
Let SWT paint the columns (except the one with the plot) to get native
indentation and abbreviation of texts.

Estimate the column width of the commit id column using "bbbbbbb"
instead of "0000000", since "0" is actually quite narrow in many fonts.
"mmmmmmm" would be the most secure variant, but that leads to much empty
space in practice, so "bbbbbbb" seems like a good compromise between
excessive free space and cutting the end often.

Bug:539553
Change-Id: Ife9a56bcfd55b4dab4728a96d7d582914a1a6853
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>