Update patch set 2
Patch Set 2:

Thomas, please test this version again and watch closely for how the first column looks like. With your suggestion almost all SHA1s were missing 1 or 2 characters on Windows (due to more indentation applied, but our width calculation assuming little indentation). Therefore I have adapted the calculation of the width to take a bigger indentation into account and to estimate the width using wider characters. Still that is just an estimation, but browsing the recent egit history there was not a single cut SHA1 in the first 3 pages. See bugzilla for a screenshot of the first column now.

Patch-set: 2