Update patch set 1
Patch Set 1:

I found a rather simple criteria now: Running the egit.core unit test suite repeatedly. The results are:
* 48-49 failing tests with mmap=true
* 19-22 failing tests with mmap=false

The fixed tests previously all reported packfile deletion failures. So changing the option back to not use mmap on Windows seems a good idea.

Regarding the question of whether the option should be available at all: Since we all seem not to know of a good reason to enable the option on Windows, I'm in favor of not providing it in the UI on Windows, and hardcoding it to false. This can easily be changed again when some user can give a concrete usage scenario, but until then it might save us some bug reports regarding locked files.

Patch-set: 1
Label: Code-Review=0