Update patch set 1
Patch Set 1:

> (1 comment)

I believe either passing the resource manager as a parameter to  HeaderText/CommitPage, or delegating the dispose()-call to HeaderText would be viable. The second option would mean that there are still two instances of the image, though. Hence I would prefer the first, if it works.

Patch-set: 1