Update patch set 1
Patch Set 1:

> To be safe, please first read the current redraw flag state and in
 > finally restore this state.

I don't think that's a good idea. The Javadoc says: "Nested calls to this method are stacked."

As long as it is ensured via try-finally that both calls are always executed, the state of drawing after both calls is exactly as it was before both calls (since it boils down to a counter, which is incremented and decremented by the implementation). Therefore if we do not call this exactly once with true and once with false, the nesting and the counter) run into the wild.

Or am I wrong with that?

Patch-set: 1