Update patch set 12
Patch Set 12:

(5 comments)

Going in the right direction. The dialog should have an image (could be UIIcons.WIZBAN_FETCH_GERRIT) and the title text and description probably should be adapted. (It says one could clone a repo, but I don't see a way to do so.) Also, the "Add..." button should either be removed, or if we keep it the dialog should propose only Gerrit-configured local repos. I'd opt for removing it for now.

In a later refactoring we might try to incorporate the GerritSelectRepositoryPage directly into the existing wizard.

Patch-set: 12
Label: Verified=0