Update patch set 2
Patch Set 2: Code-Review+2

> (1 comment)

Such comments are not meant to demotivate you. Despite the fact that we try to find problems in review before changes are submitted we obviously don't catch all of them. Thanks for working on this. If you miss keywords in the contribution guide you can add them.

Patch-set: 2
Label: Code-Review=+2