Update patch set 13
Patch Set 13:

(3 comments)

> 1. How should I go about removing the add button? I have now
 > extracted the part where the button is created into its own
 > protected method that I can just override with an empty content.
 > That doesn't seem very good. Any suggestions?

Yours is one way of doing this, but it leaves that now unused tb Container around. This doesn't need to be hyper-configurable. I had thought of a new flag in the constructor "boolean allowAdd", and then set the number of columns accordingly ("allowAdd ? 2 : 1") and just wrap the whole tb Container and creation of the add button in an "if (allowAdd) { ... }". Maybe not the super-cleanest way, but simple and fully sufficient for this use case.

Just add a new constructor, and implement the two existing constructors as "this(true, true)" and "this(allowBare, true)", respectively. Then use "super(false, false)" in the GerritSelectRepositoryPage constructor.

If we find more use cases for sub-classes of this, we can refactor that then to make this more generally configurable.

 > 2. Also, do you have suggestions for a page description? I can't
 > really seem to find a suitable solution.

You do set it to null now. Seems to be OK. It doesn't need an additional description, the title says it all already.

But the WizardDialog still has no image. It should have one; otherwise it looks strange, especially in dark theme.

Patch-set: 13