GitSelectRepositoryPage: fix broken layout
The changes made for the new GerritSelectRepositoryPage inadvertently
reordered the creation of the "Bare repositories not shown" message
Composite and of the add button composite. This results in a broken
layout when allowAdd is true.

Problem visible in the GitImportWizard when importing from an existing
local repository. Fix this be restoring the previous order of creating
these Composites.

Change-Id: I795b6d6197a47eb6c8803c1753f62bd6b6294e05
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>