Parallelize multi-repository pull
For each pulled repository a job is created that performs the pull
operation and refreshes affected projects. Progress of each pull is
reported to the job's monitor. Each pull job defines a scheduling rule
for the affected projects.
The overall action waits until all scheduled jobs are
terminated. It does not need to have a scheduling rule anymore.

Bug: 540548
Change-Id: Idc6f498197705d7b2c681dac9536470336b1ded8
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>