Update patch set 1
Patch Set 1:

I've been toying with idea at the back of my head, too. Could be done for other operations, too (Fetch, multi-repo branch switch, perhaps Push).

What happens if there are several repos with ssh remote URIs and password-protected keys with the passwords not in secure store? Will the user get multiple layered credential dialogs, or will they appear one after the other? What if some of the repos use the same password-protected key? Is it guaranteed that the user gets asked for a password for that key only once? What if the user mis-types the password?

Patch-set: 1
Label: Code-Review=0