Improve expand/collapse performance
When the number of nodes in the staging view (with tree presentation
mode) or in the tree compare view is big, then there is a very notable
repainting of the scrollbar while the tree gets expanded. Avoid this (in
all trees with potentially large content) by disabling redraw.

Note: There is still blocking (and refreshing of the mouse cursor) for
some seconds when expanding a huge tree compare view, but it surely
takes much less time than before.

Change-Id: I39021953dbb9b338b6fe013e1962daca645f871d
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>