Update patch set 5
Patch Set 5:

> Thomas, any comments from your side? Are we done, or is something
 > missing for merge?

I don't have time to review this in the near future or even to try it out. My questions about what happens with the CredentialsProvider if we have multiple repos with ssh URIs and password-protected keys etc. are still unanswered. At the very least I expect someone to try this case and report the experience.

What's the rush anyway?

Patch-set: 5