Update patch set 5
Patch Set 5:

Right, credentials. Is there a way to know if a pull will ask for a password upfront? We could group such jobs so that the run sequentially.

And yes, I would like to see it soon because I use multi pull every day. Therefore I think we should add a preference for it until we have a solution for credentials provider.

Patch-set: 5
Label: Verified=0