Have an initially selected repository in Reflog
The reflog view already tracks the workbench selection to show the
matching repository. Therefore it is reasonable to show the repository
for the current selection when opening the view.

This avoids that the user has to either first select the repository from
the toolbar or to click somewhere to explicitly set the selection.

Bug: 539799
Change-Id: I84d9ddeef9dda4867bbc359a1a399ff3508760d0
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>