Update patch set 6
Patch Set 6: Code-Review-1

This works mostly well for me.
One thing that I recognized: I changed one repository to SSH. A login dialog appeared and I cancelled the login. The process continues with the other repositories as expected. At the end, the cancelled job stays in the progress view (properly marked as cancelled) and has to be cleaned up manually. On master the process continues as well, but at the end the progress view is clean.

Patch-set: 6
Label: Code-Review=-1