Select and reveal branch node after rename
Since the active branch name may completely change its position in the
repositories view (especially with hierarchical layout active), we
should reveal the current branch again after a rename.

This code implements select and reveal for RefNodes. However, even
though the code works fine (as can be seen during debugging), the
desired effect is not achieved. It seems like yet another refresh
happens on the view, and the selection is lost again.

Bug: 539366
Change-Id: Iac3f652812ff4ad0028ea564908af2957409b5db
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>