CommitMessageViewer: more consistent order of context menu entries
The standard order for cut-copy-paste when "select all" is also
shown appears to be cut-copy-paste-select_all. For instance in the
global edit menu, or in the context menu in the problems view, or
in our own context menu in the remote configuration dialog.

Thus respect this order also in the CommitMessageViewer: move "copy"
above "select all".

Change-Id: I37dcb01386262a6c4e08ff4dd323176b0a44ad9a