Update patch set 2
Patch Set 2:

Sorry, I was not precise enough. This does not break the github project build. It only leads to errors in the IDE when all projects are in the same workspace, due to the github target still referring to the old mockito version, which is no longer available in the egit target.

Patch-set: 2
Reviewer: Gerrit User 440 <440@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0