Staging view: calculate the full title text in the background
The computation of the title of the staging view is fairly expensive.
First, computing the branch status isn't cheap, and when the repository
is in detached head state, we even scan the reflog trying to find out
whether a tag points to the commit so that we could show the tag name.
Doing all this in the UI thread is not optimal, and apparently leads to
noticeable UI freezes at least on Windows.

Improve this by leveraging the asynchronous label provider framework:
basically the title of the staging view shows the same information
as the repositories view does for a repository node. So create a
repository node and use a RepositoryTreeNodeLabelProvider to calculate
its label, then set that text as the staging view title text.

The RepositoryTreeNodeLabelProvider uses the asynchronous lightweight
decorator framework and thus computes all the expensive things in the
background and then fires a LabelProviderChangedEvent once the
decorations are ready. We re-use this same mechanism now to update the
staging view title once the decorations are ready.

Bug: 542513
Change-Id: I2b0630684d5ac983a3d241b25210c62c25c99f71
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>