[discussion] require minimum versions of core bundles
Due to the recent question whether or not we can use a Mars-only API of
StructuredSelection I remembered that Xtext now not only requires
minimum versions of features (for P2 installation), but also minimum
versions of bundles. That avoids runtime errors where users somehow
managed to use the bundles in a too old environment.

I suggest that we also increase the minimum required version for some
"core" bundles. This change exemplary does that for
* core.runtime
* jface, jface.*

The versions are taken from the resolved egit-4.5 target platform. If
this seems reasonable to other developers, I would continue with some
more core bundles.

Change-Id: Idbadca9b0dba612c1ff08177995c859aa814f93c
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>