Update patch set 1
Patch Set 1:

(1 comment)

> if GitLabelProvider.getStyledText is called with repository objects
 > from other places?

Don't see anything via "find references" on GitLabelProvider. Seems to be used mostly for branches or commits. That is not to say that we might have UI freezes because of this label provider. But for this bug (staging view title) it's of no concern.

Patch-set: 1
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0