Enable "Show Whitespace" for Java comparisons with editable index
The compare framework somehow only shows whitespace characters (if
the option is enabled) for JAva files if the content to be shown is
backed by a file.

In "Compare with index", the index version is not backed by a file.
Change EditableRevision to provide an editable IStorageEditorInput
also in this case, so that the framework does properly enable the
"show whitespace" functionality.

Handling of text changes works a bit differently then. The framework
can't save to an IStorage, so hook up the compare editor's flush
mechanism such that it updates the contents of the EditableRevision,
which will then trigger the content change listener as before, which
will update the index.

Bug: 522245
Change-Id: I2e65c68119908b605be3199c92d42d4a7e6cfce1
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>