Merge editor: enable "Show Whitespace", and fix after c73325de
Commit c73325de enhanced comparisons with an editable index.
Unfortunately the change broke the merge editor for non-index
inputs (GitMergeEditorInput, line 365: "if (!conflicting ||
useWorkspace)"). In these cases, saving didn't work anymore.
(Could occur if the merge editor was opened on a conflicting
file and the merge editor input preference was is set to
"Workspace (pre-merged by git)".)

The quick fix is to use the new mechanism introduced in commit
c73325de in all cases. This enables display of whitespace also
for these comparisons in the merge editor, and makes save work
again.

Note: with or without commit c73325de and this commit here the
merge editor is not updated properly if the resource is changed
outside the merge editor. This is bug 543495.

Bug: 522245
Change-Id: I9b2d383f3a9ff43f9f43d5f4e433b9ca88666869
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>