Make merge viewers react better on outside changes
Use the standard LocalResourceTypedElement from Platform Team for
editable items backed by an IResource instead of our own
ResourceEditableRevision. Likewise, use LocalNonWorkspaceTypedElement
for editable items backed by a non-workspace File. Make the latter
provide a FileStoreEditorInput. Using standard editor inputs as in
normal editors makes the merge editors pick up changes done elsewhere
and vice versa. Unfortunately the class needs to be a subclass of
LocalResourceTypedElement because the framework makes a number of
"instanceof" tests, but we have to override a number of more methods
to ensure that the right SharedDocumentAdapter (and byte buffer if not
connected) is used.

EditableRevision thus is only for "Compare with Index", i.e., editable
index items. LocationEditableRevision and ResourceEditableRevision
are used for conflict resolutions with "Last HEAD (unmerged)" as input.
In this case it's not possible to provide an editor input that would
match normal inputs for IResources or Files: the initial content must
come from the index stage 2, and the behavior if the (pre-merged)
working tree file is edited in parallel in another editor is confusing
anyway.

LocalFileRevision is deleted. It was originally copied from Platform
Team, but it's the wrong abstraction. It is intended for comparisons
with the local history, which are never handled by EGit, but was used in
EGit to represent an IFile in the working tree. But EGit also already
had its own WorkspaceFileRevision and WorkingTreeFileRevision for
representing items in the working tree, so use those instead. Remove
label computation code related to LocalFileRevision.

Bug: 543495
Change-Id: I933b7aa915bc9fed8f16e5bfc7b34957cec84a57
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>