Update patch set 2
Patch Set 2: -Code-Review

As far as I see, this works now well, even (within limits) for non-workspace Xtext files. (Tested with TPD.)

With Xtext, there are a few idiosyncrasies we cannot get around:

* "Show whitespace" just doesn't work for non-workspace items (non-workspace files or index/commit blobs). This is due to the implementation of Xtext's DefaultMergeViewer. Nothing we can do about it.
* While concurrent editing works nicely for text files, Xtext implements its own mechanism that is different. The user gets asked whether he wants to reload a file if it's been modified outside an Xtext editor. In the compare editor, the user gets asked upon save if there have been concurrent modifications.

There is one problem I have not figured out why it occurs: when a non-workspace Xtext file is edited and saved, the staging view does not update. This works fine for text files. I have not been able to figure out what is different in the save implementation in an Xtext editor. It _does_ work in the compare editor (for instance, compare a Xtext file (not in the workspace) from a commit with the non-workspace file in the working tree, then edit and save it in the compare editor). The work-around is to refresh the staging view explicitly. Since non-workspace Xtext files should be rare I do not consider this a problem that should block this change.

Patch-set: 2
Label: Code-Review=0