CommitFileDiffViewer: better performance for huge commits
Displaying the list of files changed might still freeze the UI for
really huge commits (> 10'000 changed files). Turns out that using a
SWT.VIRTUAL table is much faster for this use case.

Make the FileDiffLabelProvider safe against being passed null elements.
Improve getting icons for the FileDiffs; make sure that already the
base images are cached in the ResourceManager. Avoid using EGit's
DecorationOverlayDescriptor; it loads the base image's ImageData to
determine the size of the base image, and this loading is entirely
uncached!

In CommitFileDiffViewer, fix a number of performance problems with
selections:

* "select all" was very slow; instead of going through the content
  provider call selectAll() on the SWT table.
* ensure doSetItemCount(0) is equivalent to the faster doRemoveAll().
* ensure setSelection(StructuredSelection.EMPTY) goes through
  doDeselectAll().
* Avoid expensive preservingSelection() when the selection is huge by
  clearing it before setInput() if it's known that it cannot re-apply.

Bug: 544237
Change-Id: Ia4634993246e35d6d79a35f3cfa02c0e8127f2ef
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>
Also-by: Andrey Loskutov <loskutov@gmx.de>