CommitFileDiffViewer: better performance for huge commits
Displaying the list of files changed might still freeze the UI for
really huge commits (> 10'000 changed files). Turns out that using a
virtual table is much faster for this use case.

Change the table to SWT.VIRTUAL, and do the sorting in the content
provider. (Can't have a ViewerComparator with a virtual table.)

Bug: 544237
Change-Id: Ia4634993246e35d6d79a35f3cfa02c0e8127f2ef
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>