CommitFileDiffViewer: better performance for huge commits
Displaying the list of files changed might still freeze the UI for
really huge commits (> 10'000 changed files). Turns out that using a
SWT.VIRTUAL table is much faster for this use case.

Make the FileDiffLabelProvider safe against being passed null elements.
Improve getting icons for the FileDiffs; make sure that already the
base images are cached in the ResourceManager. Avoid using EGit's
DecorationOverlayDescriptor; it loads the base image's ImageData to
determine the size of the base image, and this loading is entirely
uncached!

FileDiffs are shown elsewhere, too, for instance in the fetch result
dialog. Provide an explicit WorkbenchAdapter for these cases. This
still uses the potentially expensive DecorationOverlayDescriptor, but
the separation opens the way to eventually move the decorating into
an asynchronous lightweight decorator, and it's a much less prominent
place, too.

Move the background loading out of the content provider and into
CommitFileDiffViewer; doing this in the content provider confuses
the hell out of a virtual table. Do it properly outside of any
table-related manipulations.

In CommitFileDiffViewer, fix a number of performance problems with
selections:

* "select all" was very slow; instead of going through the content
  provider call selectAll() on the SWT table.
* ensure setSelection(StructuredSelection.EMPTY) goes through
  doDeselectAll().
* Avoid expensive preservingSelection() when the selection is huge by
  clearing it before setInput() if it's known that it cannot re-apply.

Disable SWT.MULTI On GTK with SWT versions < 4919 (Eclipse 4.10).
Combined with SWT.VIRTUAL it may lead to extended UI freezes because
of bug 499850. (No such problem is known on OS X or Windows.) Provide
a new context menu action "Copy All Paths".

Bug: 544237
Change-Id: Ia4634993246e35d6d79a35f3cfa02c0e8127f2ef
Also-by: Thomas Wolf <thomas.wolf@paranor.ch>
Also-by: Andrey Loskutov <loskutov@gmx.de>
Signed-off-by: Tim Neumann <Tim.Neumann@advantest.com>