Update patch set 6
Patch Set 6:

This looks fairly good now on CentOS 7.6, GTK 3.22.30, tested with the egit-4.6 target platform and my commit with 30'000 empty files. Might still get UI freezes when there is a large selection in the viewer, though. Now off to test this on Mac.

One problem: after a "select all" in the file viewer and then selecting another repository in the repo view, the selection is removed, but it looks like there's no selection change event, and the viewer is not updated. We'll have to find a work-around for that strange behavior.

Patch-set: 6