Update patch set 2
Patch Set 2: Code-Review+1

I liked the previous version more, since I'm also used to Android Studios list of actions that are shown in an empty IDE at the very center of the application.

However, that is probably only personal taste, and in an approach of making this similar in multiple views, it seems okay.

Nevertheless, the hard line break in the project explorers text needs to be removed, and instead the label shall adapt its wrapping, just like the repositories view label does.

Patch-set: 2
Label: Code-Review=+1