Avoid excessive pruning in RepositoryCache
Inline the pruning into the other operations to avoid that we re-check
all repositories on every operation, even when we are looking only for
a specific repository. Don't call getAllRepositories() when looking
for a particular repository. Instead iterate the cached repositories
explicitly and check the existence of the git directory only for the
final candidate(s).

This drastically reduces the number of times we do check whether a
git directory exists on disk, which is a potentially expensive
operation.

Bug: 544512
Change-Id: I3a94c61261ffdd9b96fe86f86528a1daa3379256
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>