Update patch set 15
Patch Set 15:

> Patch Set 15: Code-Review-1
> 
> I've found the following problems with the current state:

As i've already wrote in the comments this change should be abandoned after integration of the code to JGit. The UI / EGit part have to be rewritten anyway for use he JGit API.

>   * While the external diff viewer is open, the eclipse UI is _completely_ frozen

The UI freeze is known, but it was ok for our developers. They use it as mergetool and during a merge conflict resolving you do not want to work with something in parallel. But this should be implemented without blocking the UI, e.g. for difftool you want to see diff in external tool and still work with Eclipse.

>   * For the files I tested with the path given to the external diff viewer is wrong (contains a double quote).

We are wotking at Windows with kdiff3, Meld, SourceGear DiffMerge, Beyond Compare 3/4 and IBM Rational Rhapsody DiffMerge. The same configuration works with "Git for Windows" in the bash shell and Eclipse as "git difftool" and "git mergetool". I'm working at Linux (Debian and Arch) at home with meld. Please give some more information about you enviroment.

Patch-set: 15