Avoid no-op checkouts during multi repository checkout
When using the multi-repository "switch branch" menu item, all selected
repositories are checked out, even if they are already on the target
branch. This has the nasty side effect of opening closed projects in
those repositories, where nothing needs to be done.

Therefore filter the list of repositories before starting the actual
work. This avoids opening the closed projects and generally leads to
less refreshes of the workspace projects.

Bug:544545
Change-Id: If215489c1bbf00e03501efeb07ca60d588e440be
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>