Rebase interactive view: consistent continue/skip/abort enablement
The rebase interactive view can also be open and show some rebase
state when a non-interactive rebase stops on a conflict. Make sure
that in this case the continue/skip/abort buttons are enabled
correctly, same as the buttons in the staging view.

Also ensure that the user cannot modify the ongoing plan in such a
case; the view doesn't allow this for interactive rebases either.

Bug: 525067
Change-Id: I7bcea7471a0e4a54736096d8b9cd2b2c90461357
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>