Shorten SHA1 in editor title
When opening a commit from history, it used the full SHA1 in the editor
title, leading to very wide tabs. Use a short version instead.

The shortening of the SHA1 was done in the editor input calculation
instead of the getContentIdentifier() method to avoid side effects at
other callers of getContentIdentifier(), which use that for equals()
comparisons of revisions.

Bug:544983
Change-Id: Ic15c12d0cd4827090d124de10765b76c62e1096e
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>