Update patch set 1
Patch Set 1:

> WTF? Why does JGit again _not_ have that canLocateSigningKey()
 > function?

Is it right that the jgit version consumed by this build is one that is created by jgit master builds after jgit merges? If so, there might be a kind of rivaling between ci and staging-ci, since we currently trigger 2 builds on every change. That could lead to one of them deploying some version you don't want.

Patch-set: 1