Update patch set 1
Patch Set 1:

> > WTF? Why does JGit again _not_ have that canLocateSigningKey()
 > > function?
 > 
 > Is it right that the jgit version consumed by this build is one
 > that is created by jgit master builds after jgit merges? If so,
 > there might be a kind of rivaling between ci and staging-ci, since
 > we currently trigger 2 builds on every change. That could lead to
 > one of them deploying some version you don't want.

I rather think the JGit stable-5.3 build publishes to the same location. Gunnar's change that introduced this method is only on master. The last JGit master build (before my retriggering today) was on March 1, but on March 3 stable-5.3 got built.[1]

To prevent this JGit should merge master to stable-5.3. Matthias is going to do so anyway this week.

[1] https://ci.eclipse.org/jgit/job/stable/job/jgit-stable/851/

Patch-set: 1