[history] Re-introduce synchronization in GenerateHistoryJob
Commit cdbe0608 had removed the synchronization on the 'loadedCommits'
SWTCommitList "since only that job accesses it". That is not quite
correct: an SWTCommitList disposes itself when the associated SWT
control (the CommitGraphTable) is disposed. This sometimes could cause
NPEs when the history view was closed while a GenerateHistoryJob was
still running.

Re-introduce the synchronization. Make sure the job accesses the
list only inside a synchronized block, and handle the list being
disposed between synchronized blocks.

Bug: 544671
Change-Id: I8e2ad3aa17ac2fb67f63f1e010389ed0bd172cbd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>