[history] Prevent NPE in GenerateHistoryJob
Commit cdbe0608 had removed the synchronization on the 'loadedCommits'
SWTCommitList "since only that job accesses it". That was not quite
correct: an SWTCommitList disposed itself when the associated SWT
control (the CommitGraphTable) was disposed. This sometimes could cause
NPEs when the history view was closed while a GenerateHistoryJob was
still running.

Remove this auto-disposal of the list. In the history page, the job
is properly shut down when the page is disposed or a new job is
started, and in the CommitSelectionDialog we can clear this list
via a DisposeListener explicitly.

Bug: 544671
Change-Id: I8e2ad3aa17ac2fb67f63f1e010389ed0bd172cbd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>