Update patch set 1
Patch Set 1:

> Your change has a side effect where I'm not sure you intended it.

I thought it was a feature. But if you prefer I can change it such that the list stays only if the text is a prefix of /refs/changes (and not empty, and not a single /), or is only digits, or "/refs/changes" is a prefix of the text.

Patch-set: 1
Label: Verified=0