Update patch set 1
Patch Set 1:

Please test it for yourself. Take the URL of this change, and after the network stuff is done, change the "refs" to "whatever" in the text. Hit Ctrl-Space and you get hundreds of suggested changes and versions, that _are completely unrelated_ to the change number in the text field. I understand that you added some less restrictive patterns in the pattern matching, but I expected that even with those new patterns the suggestions would have some correlation to the current input.

Patch-set: 1