Update patch set 1
Patch Set 1:

BTW: let Eclipse or git select the Gerrit ChangeId. If you're using the git command-line, Gerrit uses a commit hook to generate that ID, and there's a git-review helper. In Eclipse, configure the clone for Gerrit (for instance, "Gerrit Configuration..." from the context menu on the correct remote node in the Git Repositories view) and make sure you have the Gerrit button in the staging view active when you commit. EGit will fill in a dummy change-ID (all zeroes); it will be replaced when you commit (doesn't need the aforementioned hook).

Patch-set: 1