Update patch set 5
Patch Set 5:

> > > However, what probably also should be tested:  Unshared nested
 > > > project exists; parent project gets shared: both projects
 > should
 > > be
 > > > shared afterwards.
 > >
 > > Indeed, a good test-case. And I quickly did a manual check, this
 > > one fails with the current patch. Parent project cannot be
 > deleted.
 > 
 > Another test case: parent and nested project, both unshared. Nested
 > project gets shared. Parent project remains unshared. Then parent
 > project is shared. And one also needs to think about what happens
 > if the repo path is changed by the user! Consider:
 > 
 > /parent/.project
 > /parent/nested/.project
 > 
 > I share "nested" to repo path /projects, so it ends up at
 > /projects/parent/nested/.project. (Or /projects/nested/.project? I
 > actually don't know; didn't try this patch.) When I later want to
 > share "parent", am I allowed to choose a different repo path? Maybe
 > even inside nested? Or do we impose some limitations? If so, which
 > ones?

In this case, I think that we can assume that Eclipse does not keep track of this history. So, when the nested project is shared, the parent becomes a project without nested projects.
In that case, you are free to push the parent to any location you like.

A good test case would be indeed, if you are allowed to push (any project) as the parent of an already project, making the latter a nested project. Though this case might be similar as pushing to a directory that already contains contents. I don't know what would happen currently.

Patch-set: 5
Reviewer: Gerrit User 4326 <4326@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0